Enable Mixed Modulestore STUD-1540 Refactor get_parent_locations STUD-1663

+++ common/lib/xmodule/xmodule/modulestore/mixed.py
@@ -11,15 +11,14 @@ from opaque_keys import InvalidKeyError
 
 from . import ModuleStoreWriteBase
 from xmodule.modulestore.django import create_modulestore_instance, loc_mapper
-from opaque_keys.edx.locations import Location
-from xmodule.modulestore import XML_MODULESTORE_TYPE
-from opaque_keys.edx.locator import CourseLocator, Locator, BlockUsageLocator
+from opaque_keys.edx.locator import CourseLocator, BlockUsageLocator
 from xmodule.modulestore.exceptions import ItemNotFoundError
 from opaque_keys.edx.keys import CourseKey, UsageKey
 from xmodule.modulestore.mongo.base import MongoModuleStore
 from xmodule.modulestore.split_mongo.split import SplitMongoModuleStore
 from opaque_keys.edx.locations import SlashSeparatedCourseKey
 import itertools
+
 
 log = logging.getLogger(__name__)
 
@@ -31,11 +30,11 @@ class MixedModuleStore(ModuleStoreWriteBase):
     def __init__(self, mappings, stores, i18n_service=None, **kwargs):
         """
         Initialize a MixedModuleStore. Here we look into our passed in kwargs which should be a
-        collection of other modulestore configuration informations
+        collection of other modulestore configuration information
         """
         super(MixedModuleStore, self).__init__(**kwargs)
 
-        self.modulestores = {}
+        self.modulestores = []
         self.mappings = {}
 
         for course_id, store_name in mappings.iteritems():
@@ -48,36 +47,63 @@ class MixedModuleStore(ModuleStoreWriteBase):
                     log.exception("Invalid MixedModuleStore configuration. Unable to parse course_id %r", course_id)
                     continue
 
-        if 'default' not in stores:
-            raise Exception('Missing a default modulestore in the MixedModuleStore __init__ method.')
-
-        for key, store in stores.iteritems():
-            is_xml = 'XMLModuleStore' in store['ENGINE']
+        for store_settings in stores:
+            key = store_settings['NAME']
+            is_xml = 'XMLModuleStore' in store_settings['ENGINE']
             if is_xml:
                 # restrict xml to only load courses in mapping
-                store['OPTIONS']['course_ids'] = [
+                store_settings['OPTIONS']['course_ids'] = [
                     course_key.to_deprecated_string()
                     for course_key, store_key in self.mappings.iteritems()
                     if store_key == key
                 ]
-            self.modulestores[key] = create_modulestore_instance(
-                store['ENGINE'],
-                # XMLModuleStore's don't have doc store configs
-                store.get('DOC_STORE_CONFIG', {}),
-                store['OPTIONS'],
+            store = create_modulestore_instance(
+                store_settings['ENGINE'],
+                store_settings.get('DOC_STORE_CONFIG', {}),
+                store_settings.get('OPTIONS', {}),
                 i18n_service=i18n_service,
             )
+            if key == 'split':
+                store.loc_mapper = loc_mapper()
+            # replace all named pointers to the store into actual pointers
+            for course_key, store_name in self.mappings.iteritems():
+                if store_name == key:
+                    self.mappings[course_key] = store
+            self.modulestores.append(store)
 
-    def _get_modulestore_for_courseid(self, course_id):
+    def _get_modulestore_for_courseid(self, course_id=None):
         """
         For a given course_id, look in the mapping table and see if it has been pinned
         to a particular modulestore
+
+        If course_id is None, returns the first (ordered) store as the default
         """
-        # TODO when this becomes a router capable of handling more than one r/w backend
-        # we'll need to generalize this to handle mappings from old Locations w/o full
-        # course_id in much the same way as loc_mapper().translate_location does.
-        mapping = self.mappings.get(course_id, 'default')
-        return self.modulestores[mapping]
+        if course_id is not None:
+            mapping = self.mappings.get(course_id, None)
+            if mapping is not None:
+                return mapping
+            else:
+                for store in self.modulestores:
+                    if isinstance(course_id, store.reference_type) and store.has_course(course_id):
+                        self.mappings[course_id] = store
+                        return store
+
+        # return the first store, as the default
+        return self.modulestores[0]
+
+    def _get_modulestore_by_type(self, modulestore_type):
+        """
+        This method should only really be used by tests and migration scripts when necessary.
+        Returns the module store as requested by type.  The type can be:
+
+            SPLIT_MONGO_MODULESTORE_TYPE
+            MONGO_MODULESTORE_TYPE
+            XML_MODULESTORE_TYPE
+        """
+        for store in self.modulestores:
+            if store.get_modulestore_type() == modulestore_type:
+                return store
+        return None
 
     def has_item(self, usage_key):
         """
@@ -129,33 +155,23 @@ class MixedModuleStore(ModuleStoreWriteBase):
         Returns a list containing the top level XModuleDescriptors of the courses
         in this modulestore.
         '''
-        # order the modulestores and ensure no dupes (default may be a dupe of a named store)
-        # remove 'draft' as we know it's a functional dupe of 'direct' (ugly hardcoding)
-        stores = set([value for key, value in self.modulestores.iteritems() if key != 'draft'])
-        stores = sorted(stores, cmp=_compare_stores)
-
-        courses = {}  # a dictionary of stringified course locations to course objects
-        has_locators = any(issubclass(CourseLocator, store.reference_type) for store in stores)
-        for store in stores:
+        courses = {}  # a dictionary of course keys to course objects
+        has_locators = any(issubclass(CourseLocator, store.reference_type) for store in self.modulestores)
+        for store in self.modulestores:
             store_courses = store.get_courses()
             # filter out ones which were fetched from earlier stores but locations may not be ==
             for course in store_courses:
-                course_location = unicode(course.location)
-                if course_location not in courses:
-                    if has_locators and isinstance(course.location, Location):
+                if course.id not in courses:
+                    if has_locators and isinstance(course.id, SlashSeparatedCourseKey):
                         # see if a locator version of course is in the result
                         try:
-                            # if there's no existing mapping, then the course can't have been in split
-                            course_locator = loc_mapper().translate_location(
-                                course.location,
-                                add_entry_if_missing=False
-                            )
-                            if unicode(course_locator) not in courses:
-                                courses[course_location] = course
+                            course_locator = loc_mapper().translate_location_to_course_locator(course.id)
+                            if course_locator in courses:
+                                continue
                         except ItemNotFoundError:
-                            courses[course_location] = course
-                    else:
-                        courses[course_location] = course
+                            # if there's no existing mapping, then the course can't have been in split
+                            pass
+                    courses[course.id] = course
 
         return courses.values()
 
@@ -175,11 +191,13 @@ class MixedModuleStore(ModuleStoreWriteBase):
 
     def has_course(self, course_id, ignore_case=False):
         """
-        returns whether the course exists
+        returns the course_id of the course if it was found, else None
+        Note: we return the course_id instead of a boolean here since the found course may have
+           a different id than the given course_id when ignore_case is True.
 
         Args:
         * course_id (CourseKey)
-        * ignore_case (bool): Tf True, do a case insensitive search. If
+        * ignore_case (bool): If True, do a case insensitive search. If
             False, do a case sensitive search
         """
         assert(isinstance(course_id, CourseKey))
@@ -192,14 +210,17 @@ class MixedModuleStore(ModuleStoreWriteBase):
         """
         assert(isinstance(course_key, CourseKey))
         store = self._get_modulestore_for_courseid(course_key)
-        return store.delete_course(course_key, user_id)
+        if hasattr(store, 'delete_course'):
+            return store.delete_course(course_key, user_id)
+        else:
+            raise NotImplementedError(u"Cannot delete a course on store {}".format(store))
 
-    def get_parent_locations(self, location):
+    def get_parent_locations(self, location, **kwargs):
         """
         returns the parent locations for a given location
         """
         store = self._get_modulestore_for_courseid(location.course_key)
-        return store.get_parent_locations(location)
+        return store.get_parent_locations(location, **kwargs)
 
     def get_modulestore_type(self, course_id):
         """
@@ -209,7 +230,7 @@ class MixedModuleStore(ModuleStoreWriteBase):
         "mongo" for old-style MongoDB backed courses,
         "split" for new-style split MongoDB backed courses.
         """
-        return self._get_modulestore_for_courseid(course_id).get_modulestore_type(course_id)
+        return self._get_modulestore_for_courseid(course_id).get_modulestore_type()
 
     def get_orphans(self, course_key):
         """
@@ -226,11 +247,11 @@ class MixedModuleStore(ModuleStoreWriteBase):
         course_dir where course loading failed.
         """
         errs = {}
-        for store in self.modulestores.values():
+        for store in self.modulestores:
             errs.update(store.get_errored_courses())
         return errs
 
-    def create_course(self, org, offering, user_id=None, fields=None, store_name='default', **kwargs):
+    def create_course(self, org, offering, user_id=None, fields=None, **kwargs):
         """
         Creates and returns the course.
 
@@ -239,15 +260,14 @@ class MixedModuleStore(ModuleStoreWriteBase):
             offering (str): the name of the course offering
             user_id: id of the user creating the course
             fields (dict): Fields to set on the course at initialization
-            store_name (str): the name of the modulestore that we will create this course within
             kwargs: Any optional arguments understood by a subset of modulestores to customize instantiation
 
         Returns: a CourseDescriptor
         """
-        store = self.modulestores[store_name]
+        store = self._get_modulestore_for_courseid(None)
 
         if not hasattr(store, 'create_course'):
-            raise NotImplementedError(u"Cannot create a course on store %s" % store_name)
+            raise NotImplementedError(u"Cannot create a course on store {}".format(store))
 
         return store.create_course(org, offering, user_id, fields, **kwargs)
 
@@ -273,7 +293,7 @@ class MixedModuleStore(ModuleStoreWriteBase):
             if location is None:
                 location = course_id.make_usage_key(category, block_id)
             # do the actual creation
-            xblock = store.create_and_save_xmodule(location, **kwargs)
+            xblock = self.create_and_save_xmodule(location, user_id, **kwargs)
             # don't forget to attach to parent
             if parent_loc is not None and not 'detached' in xblock._class_tags:
                 parent = store.get_item(parent_loc)
@@ -302,7 +322,10 @@ class MixedModuleStore(ModuleStoreWriteBase):
         """
         course_id = xblock.scope_ids.usage_id.course_key
         store = self._get_modulestore_for_courseid(course_id)
-        return store.update_item(xblock, user_id)
+        if hasattr(store, 'update_item'):
+            return store.update_item(xblock, user_id, allow_not_found)
+        else:
+            raise NotImplementedError(u"Cannot update item on store {}".format(store))
 
     def delete_item(self, location, user_id=None, **kwargs):
         """
@@ -310,17 +333,31 @@ class MixedModuleStore(ModuleStoreWriteBase):
         """
         course_id = location.course_key
         store = self._get_modulestore_for_courseid(course_id)
-        return store.delete_item(location, user_id=user_id, **kwargs)
+        if hasattr(store, 'delete_item'):
+            return store.delete_item(location, user_id=user_id, **kwargs)
+        else:
+            raise NotImplementedError(u"Cannot delete item on store {}".format(store))
 
     def close_all_connections(self):
         """
         Close all db connections
         """
-        for mstore in self.modulestores.itervalues():
+        for mstore in self.modulestores:
             if hasattr(mstore, 'database'):
                 mstore.database.connection.close()
             elif hasattr(mstore, 'db'):
                 mstore.db.connection.close()
+
+    def create_xmodule(self, location, definition_data=None, metadata=None, system=None, fields={}):
+        """
+        Create the new xmodule but don't save it. Returns the new module.
+        """
+        course_id = location.course_key
+        store = self._get_modulestore_for_courseid(course_id)
+        if hasattr(store, 'create_xmodule'):
+            return store.create_xmodule(location, definition_data, metadata, system, fields)
+        else:
+            raise NotImplementedError(u"Cannot create_xmodule on store {}".format(store))
 
     def get_courses_for_wiki(self, wiki_slug):
         """
@@ -329,7 +366,7 @@ class MixedModuleStore(ModuleStoreWriteBase):
         :return: list of course locations
         """
         courses = []
-        for modulestore in self.modulestores.values():
+        for modulestore in self.modulestores:
             courses.extend(modulestore.get_courses_for_wiki(wiki_slug))
         return courses
 
@@ -345,29 +382,59 @@ class MixedModuleStore(ModuleStoreWriteBase):
             )
         )
 
+    def compute_publish_state(self, xblock):
+        """
+        Returns whether this xblock is 'draft', 'public', or 'private'.
 
-def _compare_stores(left, right):
-    """
-    Order stores via precedence: if a course is found in an earlier store, it shadows the later store.
-
-    xml stores take precedence b/c they only contain hardcoded mappings, then Locator-based ones,
-    then others. Locators before Locations because if some courses may be in both,
-    the ones in the Locator-based stores shadow the others.
-    """
-    if left.get_modulestore_type(None) == XML_MODULESTORE_TYPE:
-        if right.get_modulestore_type(None) == XML_MODULESTORE_TYPE:
-            return 0
+        'draft' content is in the process of being edited, but still has a previous
+            version visible in the LMS
+        'public' content is locked and visible in the LMS
+        'private' content is editable and not visible in the LMS
+        """
+        course_id = xblock.scope_ids.usage_id.course_key
+        store = self._get_modulestore_for_courseid(course_id)
+        if hasattr(store, 'compute_publish_state'):
+            return store.compute_publish_state(xblock)
+        elif hasattr(store, 'publish'):
+            raise NotImplementedError(u"Cannot compute_publish_state on store {}".format(store))
         else:
-            return -1
-    elif right.get_modulestore_type(None) == XML_MODULESTORE_TYPE:
-        return 1
+            # read-only store; so, everything's public
+            return 'public'
 
-    if issubclass(left.reference_type, Locator):
-        if issubclass(right.reference_type, Locator):
-            return 0
+    def publish(self, location, user_id):
+        """
+        Save a current draft to the underlying modulestore
+        Returns the newly published item.
+        """
+        course_id = location.course_key
+        store = self._get_modulestore_for_courseid(course_id)
+        if hasattr(store, 'publish'):
+            return store.publish(location, user_id)
         else:
-            return -1
-    elif issubclass(right.reference_type, Locator):
-        return 1
+            raise NotImplementedError(u"Cannot publish on store {}".format(store))
 
-    return 0
+    def unpublish(self, location, user_id):
+        """
+        Save a current draft to the underlying modulestore
+        Returns the newly unpublished item.
+        """
+        course_id = location.course_key
+        store = self._get_modulestore_for_courseid(course_id)
+        if hasattr(store, 'unpublish'):
+            return store.unpublish(location, user_id)
+        else:
+            raise NotImplementedError(u"Cannot unpublish on store {}".format(store))
+
+    def convert_to_draft(self, location, user_id):
+        """
+        Create a copy of the source and mark its revision as draft.
+
+        :param source: the location of the source (its revision must be None)
+        """
+        course_id = location.course_key
+        store = self._get_modulestore_for_courseid(course_id)
+        if hasattr(store, 'convert_to_draft'):
+            return store.convert_to_draft(location, user_id)
+        else:
+            raise NotImplementedError(u"Cannot convert_to_draft on store {}".format(store))
+
